Remove single accessor FFI implementations
They didn't seem worth it with the only application I know of which
uses them, Mesecons. They could even slow it down. These could be
re-added at a later date if my assessment is wrong, but I think it
will be better to just optimize single accessors without using FFI.

There are still some traces of the changes, including the tests and
the noexcept specifiers in voxel.h, but these are find to leave be.